Removed old editor & moved stamps
