+++ src/main/java/io/wollinger/snipsniper/sceditor/stamps/CircleStamp.java
@@ -1,4 +1,4 @@
-package io.wollinger.snipsniper.editorwindow.stamps;
+package io.wollinger.snipsniper.sceditor.stamps;
 
 import io.wollinger.snipsniper.Config;
 import io.wollinger.snipsniper.utils.InputContainer;
